+++ packages/core/src/theme/components/EditLink/index.scss
@@ -1,11 +1,13 @@
-.rp-edit-link {
-  color: var(--rp-c-brand);
-  text-decoration: none;
-  font-size: 15px;
-  font-weight: 500;
-  transition: all 0.2s ease-in-out;
+.rp-doc-footer {
+  .rp-edit-link {
+    color: var(--rp-c-brand);
+    text-decoration: none;
+    font-size: 15px;
+    font-weight: 500;
+    transition: all 0.2s ease-in-out;
 
-  &:hover {
-    color: var(--rp-c-brand-dark);
+    &:hover {
+      color: var(--rp-c-brand-dark);
+    }
   }
 }
