+++ packages/core/src/theme/components/Llms/LlmsOpenRow.scss
@@ -1,0 +1,10 @@
+.rp-outline__open-in-wrapper {
+  position: relative;
+
+  .rp-llms-view-options__menu {
+    bottom: calc(100% + 6px);
+    top: auto;
+    left: 0;
+    transform-origin: bottom left;
+  }
+}
